51 auth

+++ app/lib/queries.ts
@@ -1,12 +1,20 @@
 'use server';
 
 import prisma from '@/app/lib/prisma';
+import { checkUser } from '@/app/actions/actions';
+import { redirect } from 'next/navigation';
 
-export async function getPathwaysByEmail() {
+export async function getUserPathways() {
+  const user = await checkUser();
+
+  if (!user) {
+    redirect('/login');
+  }
+
   try {
     return await prisma.person.findFirst({
       where: {
-        email: process.env.TEST_USER_EMAIL,
+        id: user.id,
       },
       include: {
         pathways: {
@@ -26,11 +34,17 @@ export async function getPathwaysByEmail() {
   }
 }
 
-export async function getPathwayByUserId(userId: string, pathwayId: string) {
+export async function getSingleUserPathway(pathwayId: string) {
+  const user = await checkUser();
+
+  if (!user) {
+    redirect('/login');
+  }
+
   try {
     return await prisma.person.findFirst({
       where: {
-        id: userId,
+        id: user.id,
       },
       include: {
         pathways: {
@@ -53,7 +67,13 @@ export async function getPathwayByUserId(userId: string, pathwayId: string) {
   }
 }
 
-export async function getCompetency(userId: string, competencyId: string) {
+export async function getUserCompetency(competencyId: string) {
+  const user = await checkUser();
+
+  if (!user) {
+    redirect('/login');
+  }
+
   try {
     return await prisma.competency.findFirst({
       where: {
@@ -62,7 +82,7 @@ export async function getCompetency(userId: string, competencyId: string) {
       include: {
         proofs: {
           where: {
-            personId: userId,
+            personId: user.id,
           },
         },
         contentAreas: true,
@@ -74,19 +94,24 @@ export async function getCompetency(userId: string, competencyId: string) {
 }
 
 export async function createProof(data: {
-  userId: string;
   title: string;
   description: string;
   justification: string;
   competencyId: string;
 }) {
+  const user = await checkUser();
+
+  if (!user) {
+    redirect('/login');
+  }
+
   try {
     return await prisma.proof.create({
       data: {
         title: data.title,
         description: data.description,
         justification: data.justification,
-        personId: data.userId,
+        personId: user.id,
         competencyId: parseInt(data.competencyId),
       },
     });
@@ -96,10 +121,17 @@ export async function createProof(data: {
 }
 
 export async function deleteProof(proofId: number) {
+  const user = await checkUser();
+
+  if (!user) {
+    redirect('/login');
+  }
+
   try {
     return await prisma.proof.delete({
       where: {
         id: proofId,
+        personId: user.id,
       },
     });
   } catch (error) {
@@ -107,12 +139,55 @@ export async function deleteProof(proofId: number) {
   }
 }
 
-export async function getProof( proof: string) {
-  // read only for profile, so no need userid.
+export async function getProof(proof: string) {
+  // read only for profile, so no need for auth.
   try {
     return await prisma.proof.findFirst({
       where: {
         id: parseInt(proof),
+      },
+    });
+  } catch (error) {
+    console.error(error);
+  }
+}
+
+export async function createUser({
+  username,
+  hashedPassword,
+  email,
+  firstName,
+  lastName,
+  bio,
+}: {
+  username: string;
+  hashedPassword: string;
+  email: string;
+  firstName: string;
+  lastName: string;
+  bio: string;
+}) {  
+  try {
+    return await prisma.person.create({
+      data: {
+        username: username,
+        hashedPassword: hashedPassword,
+        email,
+        firstName,
+        lastName,
+        bio,
+      },
+    });
+  } catch (error) {
+    console.error(error);
+  }
+}
+
+export async function getUser(username: string) {
+  try {
+    return await prisma.person.findUnique({
+      where: {
+        username: username,
       },
     });
   } catch (error) {
